:sparkles: Correct second OECD working paper
As previously, attempting to change markup to correct structure
that Grobid put there, but which was either misplaced or not complete
or not the right one.

Still, any open questions are marked up using XML comments starting
with `FIXME`. Some parts of pages are missing or misplaced.

A lot of paragraphs were not correctly marked up, i.e. Grobid combined
them, when they were actually different paragraphs.

Another frequent modification were references to figures that were
numbered like `4.1`. Often, the second part, in this example `.1` was
placed outside the `ref` marker.